+++ src/helpers/Db.php
@@ -1026,30 +1026,27 @@ class Db
     }
 
     /**
-     * This method attempts to see if timezone data has been populated on the
-     * database server Craft is configured to use.
+     * Returns whether the database supports time zone conversions.
      *
-     * https://dev.mysql.com/doc/refman/5.7/en/time-zone-support.html
+     * This could return `false` if MySQL’s time zone tables haven’t been populated yet. See
+     * https://dev.mysql.com/doc/refman/5.7/en/time-zone-support.html for more info.
+     *
      * @param Connection|null $db
-     * @return bool `false` if this is MySQL and it appears timezone data is not available, `true` otherwise.
-     * @throws DbException
+     * @return bool
      * @since 3.6.16
      */
-    public static function validateDatabaseTimezoneSupport(?Connection $db = null): bool
+    public static function supportsTimeZones(?Connection $db = null): bool
     {
         if ($db === null) {
             $db = self::db();
         }
 
-        if ($db->getIsMysql()) {
-            $result = $db->createCommand("SELECT CONVERT_TZ('2007-03-11 2:00:00','US/Eastern','US/Central') AS time1")->queryScalar();
-
-            if (!$result) {
-                return false;
-            }
+        if ($db->getIsPgsql()) {
+            return true;
         }
 
-        return true;
+        $result = $db->createCommand("SELECT CONVERT_TZ('2007-03-11 2:00:00','US/Eastern','US/Central') AS time1")->queryScalar();
+        return (bool)$result;
     }
 
     /**